src/loadgenerator/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
blinker==1.6.2
    # via flask
brotli==1.0.9
    # via geventhttpclient
certifi==2023.5.7
    # via
    #   geventhttpclient
    #   requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via flask
configargparse==1.5.3
    # via locust
flask==2.3.2
    # via
    #   flask-basicauth
    #   flask-cors
    #   locust
flask-basicauth==0.2.0
    # via locust
flask-cors==3.0.10
    # via locust
gevent==22.10.2
    # via
    #   geventhttpclient
    #   locust
geventhttpclient==2.0.9
    # via locust
greenlet==2.0.2
    # via gevent
idna==3.4
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
locust==2.15.1
    # via -r requirements.in
grpcio-tools==1.56.0
grpc-interceptor==0.15.2
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
msgpack==1.0.5
    # via locust
psutil==5.9.5
    # via locust
pyzmq==25.1.0
    # via locust
requests==2.31.0
    # via locust
roundrobin==0.0.4
    # via locust
six==1.16.0
    # via
    #   flask-cors
    #   geventhttpclient
typing-extensions==4.6.3
    # via locust
urllib3==2.0.2
    # via requests
werkzeug==2.3.6
    # via
    #   flask
    #   locust
zope-event==4.6
    # via gevent
zope-interface==6.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
